% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\docType{data}
\name{example}
\alias{example}
\title{Example FASTA dataset}
\format{
a character vector of lenght 322
}
\source{
Simulated data with giveExampleFasta function created by Arthur Passos
}
\usage{
example
}
\description{
a small FASTA dataset with 100 sequences, included for demonstrate
for purpose this package
}
\keyword{datasets}
