% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastq2fasta.R
\name{fastq2fasta}
\alias{fastq2fasta}
\title{Converter arquivo FASTQ para FASTA}
\usage{
fastq2fasta(file, output_name = NULL)
}
\arguments{
\item{file}{Caminho para o arquivo FASTQ de entrada.}

\item{output_name}{Nome do arquivo FASTA de saída.
Se NULL ou vazio, será gerado automaticamente como `"arquivo_converted.fasta"`.}
}
\value{
Invisivelmente retorna o nome do arquivo gerado.
}
\description{
Esta função lê um arquivo FASTQ e gera um arquivo FASTA com os mesmos IDs
e sequências. O nome de saída pode ser definido pelo usuário ou gerado
automaticamente a partir do nome do arquivo de entrada.
}
\examples{
\dontrun{
fastq2fasta("reads.fastq")
fastq2fasta("reads.fastq", output_name = "dados.fasta")
}
}
