% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telomeres_hits_example.R
\docType{data}
\name{telomere_hits_example}
\alias{telomere_hits_example}
\title{Example FASTA dataset}
\format{
a character vector of lenght 322
}
\source{
Simulated data with giveExampleFasta function created by Arthur Passos
}
\usage{
telomere_hits_example
}
\description{
a small FASTA dataset with hits from example.fasta, included for demonstrate
for purpose this package
}
\keyword{datasets}
