% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telomeresfinder.R
\name{Telomeresfinder}
\alias{Telomeresfinder}
\title{Identificar sequências contendo repetições teloméricas canônicas}
\usage{
Telomeresfinder(
  fasta_file,
  telomere_repeat = NULL,
  min_repeats = NULL,
  output_fasta = "telomere_hits.fasta"
)
}
\arguments{
\item{fasta_file}{Caminho para o arquivo FASTA.}

\item{telomere_repeat}{Sequência repetitiva telomérica canônica
(exemplo: `"TTAGGG"`).}

\item{min_repeats}{Número mínimo de repetições consecutivas necessárias
para considerar a sequência como contendo telômero.}
}
\value{
Um objeto `DNAStringSet` contendo apenas as sequências que atendem ao critério.
}
\description{
Esta função busca repetições teloméricas canônicas em arquivos FASTA e
retorna apenas as sequências que possuem um número mínimo de repetições
consecutivas. Ela também considera automaticamente as versões reversa,
complementar e reverso-complementar.
}
\examples{
\dontrun{
hits <- Telomeresfinder("reads.fasta", telomere_repeat = "TTAGGG", min_repeats = 3)
}

}
