% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastq2fasta.R
\name{fastq2fasta}
\alias{fastq2fasta}
\title{Convert FASTQ file to FASTA}
\usage{
fastq2fasta(file, output_name = NULL)
}
\arguments{
\item{file}{Path to the input FASTQ file.}

\item{output_name}{Name of the output FASTA file.
If NULL or empty, it will be automatically generated as `"file_converted.fasta"`.}
}
\value{
Invisibly returns the name of the generated file.
}
\description{
This function reads a FASTQ file and generates a FASTA file with the same
IDs and sequences. The output filename can be defined by the user or
automatically generated from the input filename.
}
\examples{
\dontrun{
fastq2fasta("reads.fastq")
fastq2fasta("reads.fastq", output_name = "data.fasta")
}

}
