% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telomeresfinder.R
\name{Telomeresfinder}
\alias{Telomeresfinder}
\title{Identify sequences containing canonical telomeric repeats}
\usage{
Telomeresfinder(
  fasta_file,
  telomere_repeat = NULL,
  min_repeats = NULL,
  output_fasta = "telomere_hits.fasta"
)
}
\arguments{
\item{fasta_file}{Path to the FASTA file.}

\item{telomere_repeat}{Canonical telomeric repeat sequence
(example: `"TTAGGG"`).}

\item{min_repeats}{Minimum number of consecutive repeats required
for the sequence to be considered as containing telomere repeats.}

\item{output_fasta}{Optional path to the output FASTA file.}
}
\value{
A `DNAStringSet` object containing only the sequences that meet the criteria.
}
\description{
This function searches for canonical telomeric repeats in FASTA files
and returns only the sequences that contain a minimum number of consecutive
repeat units. It also automatically considers the reverse, complement,
and reverse-complement versions of the repeat.
}
\examples{
\dontrun{
hits <- Telomeresfinder("reads.fasta", telomere_repeat = "TTAGGG", min_repeats = 3)
}

}
